change to Long for version primitives (#79)
(cherry picked from commit bcdec530a00224eaa537eb7f20361a44040b79de)

+++ app/src/main/java/cc/sovellus/vrcaa/api/vrchat/models/Avatar.kt
@@ -37,5 +37,5 @@ data class Avatar(
     @SerializedName("updated_at")
     var updatedAt: String = "",
     @SerializedName("version")
-    var version: Int = 0
+    var version: Long = 0
 )
